python-pytest/requirements.txt
pytest>=5.0.0
allure-pytest>=2.8.18
pyflakes>=2.2.0
flake8>=3.8.3